[issue-33569] Show User Events on dedicated tab on Client-/User-Details (#33574)
Fixes #33569
